==> 06_cleanup.ipynb <==
## Clean-up
#### Some general examples of how to clean-up your SageMaker environment

import boto3

sm = boto3.Session().client(service_name="sagemaker")

#### List all my SageMaker Models 

models = []
results = sm.list_models()
while True:

    if len(results['Models']):
        models.extend(results['Models'])
    
    if 'NextToken' in results.keys():
        results = sm.list_models(NextToken=results['NextToken'])
    else:
        break
        

#### Delete all but the most recent 20 of my SageMaker Models

for m in models[20:]:
    response = sm.delete_model(ModelName=m['ModelName'])
    print(response)
